+++ Assets/Adjust/Test/TestLibraryAndroid.cs
@@ -12,14 +12,14 @@ namespace com.adjust.sdk.test
             CommandExecutor commandExecutor = new CommandExecutor(this, baseUrl, gdprUrl);
             onCommandReceivedListener = new CommandListenerAndroid(commandExecutor);
             ajoTestLibrary = new AndroidJavaObject(
-                "com.adjust.testlibrary.TestLibrary",
+                "com.adjust.test.TestLibrary",
                 baseUrl,
                 onCommandReceivedListener);
         }
 
         public void StartTestSession()
         {
-            ajoTestLibrary.Call("startTestSession", TestApp.CLIENT_SDK);
+            ajoTestLibrary.Call("startTestSession", Adjust.getSdkVersion());
         }
 
         public void AddInfoToSend(string key, string paramValue)