+++ Assets/Adjust/Windows/AdjustWindows.cs
@@ -16,7 +16,7 @@ namespace com.adjust.sdk
 {
     public class AdjustWindows
     {
-        private const string sdkPrefix = "unity4.15.0";
+        private const string sdkPrefix = "unity4.17.0";
         private static bool appLaunched = false;
 
         public static void Start(AdjustConfig adjustConfig)
@@ -228,6 +228,11 @@ namespace com.adjust.sdk
         public static string GetAdid()
         {
             return AdjustWinInterface.GetAdid();
+        }
+
+        public static string GetSdkVersion()
+        {
+            return sdkPrefix + "@" + AdjustWinInterface.GetSdkVersion();
         }
 
         public static AdjustAttribution GetAttribution()