+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+### Version 4.17.0 (13th December 2018)
+#### Added
+- Added `getSdkVersion()` method to `Adjust` interface to obtain current SDK version string.
+- Added `AdjustImei` Unity plugin to enable reading of IMEI and MEID identifiers for apps outside of Google Play Store which need this feature.
+
+#### Native SDKs
+- [iOS@v4.17.1][ios_sdk_v4.17.1]
+- [Android@v4.17.0][android_sdk_v4.17.0]
+- [Windows@v4.17.0][windows_sdk_v4.17.0]
+
+---
+
 ### Version 4.15.0 (21st September 2018)
 #### Added
 - Added `setCallbackId` method on `AdjustEvent` object for users to set custom ID on event object which will later be reported in event success/failure callbacks.
@@ -529,6 +541,7 @@
 [ios_sdk_v4.14.0]: https://github.com/adjust/ios_sdk/tree/v4.14.0
 [ios_sdk_v4.14.1]: https://github.com/adjust/ios_sdk/tree/v4.14.1
 [ios_sdk_v4.15.0]: https://github.com/adjust/ios_sdk/tree/v4.15.0
+[ios_sdk_v4.17.1]: https://github.com/adjust/ios_sdk/tree/v4.17.1
 
 [android_sdk_v3.5.0]: https://github.com/adjust/android_sdk/tree/v3.5.0
 [android_sdk_v4.1.0]: https://github.com/adjust/android_sdk/tree/v4.1.0
@@ -547,8 +560,10 @@
 [android_sdk_v4.13.0]: https://github.com/adjust/android_sdk/tree/v4.13.0
 [android_sdk_v4.14.0]: https://github.com/adjust/android_sdk/tree/v4.14.0
 [android_sdk_v4.15.1]: https://github.com/adjust/android_sdk/tree/v4.15.1
+[android_sdk_v4.17.0]: https://github.com/adjust/android_sdk/tree/v4.17.0
 
 [windows_sdk_v4.12.0]: https://github.com/adjust/windows_sdk/tree/v4.12.0
 [windows_sdk_v4.13.0]: https://github.com/adjust/windows_sdk/tree/v4.13.0
 [windows_sdk_v4.14.0]: https://github.com/adjust/windows_sdk/tree/v4.14.0
 [windows_sdk_v4.15.0]: https://github.com/adjust/windows_sdk/tree/v4.15.0
+[windows_sdk_v4.17.0]: https://github.com/adjust/windows_sdk/tree/v4.17.0